+++ app/tng/tnglora_cert_def_4_device.c
@@ -193,8 +193,8 @@ SHARED_LIB_EXPORT const atcacert_def_t g_tnglora_cert_def_4_device = {
         }
     },
     .cert_elements       = g_tnglora_cert_elements_4_device,
-    .cert_elements_count = sizeof(g_tnglora_cert_elements_4_device) / sizeof(g_tnglora_cert_elements_4_device[0]),
+    .cert_elements_count = (uint8_t)(sizeof(g_tnglora_cert_elements_4_device) / sizeof(g_tnglora_cert_elements_4_device[0])),
     .cert_template       = g_tnglora_cert_template_4_device,
-    .cert_template_size  = sizeof(g_tnglora_cert_template_4_device),
+    .cert_template_size  = (uint16_t)(sizeof(g_tnglora_cert_template_4_device)),
     .ca_cert_def         = &g_tnglora_cert_def_1_signer
 };
